% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalmodel_misc.R
\name{days_since_last_match}
\alias{days_since_last_match}
\title{Time since last match}
\usage{
days_since_last_match(team1, team2, dates, first_val = NA)
}
\arguments{
\item{team1}{Vector of team names.}

\item{team2}{Vector of team names.}

\item{dates}{a vector of dates (of type Date).}

\item{first_val}{the value used for a team's first match in the data, when there
are no preceding matches. Default is NA.}
}
\value{
A two-column matrix with the number of days since last time
each team played.
}
\description{
This function calculates the number of days since the two teams
played their respective previous matches.
}
\seealso{
\code{\link{matches_last_xdays}}
}
