% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{dDCP}
\alias{dDCP}
\title{The Dixon-Coles-Poisson Distribution}
\usage{
dDCP(x1, x2, lambda1, lambda2, rho = NULL)
}
\arguments{
\item{x1}{non-negative integer.}

\item{x2}{non-negative integer.}

\item{lambda1}{vector of non-negative parameters of the Dixon-Coles-Poisson distribution.}

\item{lambda2}{vector of non-negative parameters of the Dixon-Coles-Poisson distribution.}

\item{rho}{vector of parameters of the Dixon-Coles-Poisson distribution.}
}
\description{
Probability mass function for the Dixon-Coles-Poisson distribution.
}
\section{References}{

\itemize{
 \item{Mark J.  Dixon & Stuart G. Coles (1997) Modelling Association Football Scores and Inefficiencies in the Football Betting Market }
}
}

