% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalmodel_misc.R
\name{expg_from_probabilities}
\alias{expg_from_probabilities}
\title{Estimate the expected goals from win-draw-lose probabilities.}
\usage{
expg_from_probabilities(probabilities, rho = 0, uprx = 75)
}
\arguments{
\item{probabilities}{A 3-column matrix of win-draw-lose probabilities.}

\item{rho}{numeric. Value for the Dixon-Coles adjustment parameter. Default is 0, which is the same as no adjustment.}

\item{uprx}{numeric. The upper limit for evaluating the poisson distribution.}
}
\value{
A list with two elements. The first, expg, is a two-column matrix of
expected goals. The second, sq_error, is a numeric vector indicating the how well
the expected goals matches the probabilities using the poisson distribution.
}
\description{
This function converts outcome probabilities into expected goals,
assuming an underlying Poisson distribution, or a Dixon-Coles-Poisson
distribution with known dependence paramater rho.
}
