% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalmodel_misc.R
\name{league_table}
\alias{league_table}
\title{Create league table}
\usage{
league_table(goals1, goals2, team1, team2)
}
\arguments{
\item{goals1}{Numeric, non-negative integer. The number of goals scored by team 1.}

\item{goals2}{Numeric, non-negative integer. The number of goals scored by team 2.}

\item{team1}{Vector of team names.}

\item{team2}{Vector of team names.

#' @return
 A data.frame with one row for each team in the data.}
}
\description{
Creates a league table with basic descriptive statistics and number of points.
}
\details{
Points are calculated as 3 points for win, 1 for draw, and 0 for loss. Results
are sorted by 1) points, 2) goal difference and 3) goals scored.

Games with missing values will be skipped.
}
