% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalmodel_misc.R
\name{matches_last_xdays}
\alias{matches_last_xdays}
\title{Number of matches played in the preceding period}
\usage{
matches_last_xdays(team1, team2, dates, days_since = 30, first_val = NA)
}
\arguments{
\item{team1}{Vector of team names.}

\item{team2}{Vector of team names.}

\item{dates}{a vector of dates (of type Date).}

\item{days_since}{The number of days back in time to count matches.}

\item{first_val}{the value used for a team's first match in the data, when there
are no preceding matches. Default is NA.}
}
\value{
A two-column matrix with the number of matches played in the preceding
period of time, as given by the days_since argument.
}
\description{
This function calculates the number of matches the two teams
have played the last x days.
}
\seealso{
\code{\link{days_since_last_match}}
}
