% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalmodel_predict.R
\name{pbtts}
\alias{pbtts}
\title{Compute both-teams-to-score probabilities from expected goals.}
\usage{
pbtts(expg1, expg2, model = "poisson", dispersion = NULL, rho = NULL)
}
\arguments{
\item{expg1}{Non-negative numeric. The expected number of goals.}

\item{expg2}{Non-negative numeric. The expected number of goals.}

\item{model}{String indicating whether the goals follow a 'poisson' model (default), a Negative Binomial ('negbin'), or Conway-Maxwell-Poisson ('cmp') model.}

\item{dispersion}{Non-negative numeric. The dispersion parameter in the Negative Binomial model or the Conway-Maxwell-Poisson model.}

\item{rho}{Numeric. The Dixon-Coles adjustment.}
}
\value{
A numeric vector with probabilities for both teams to score.
}
\description{
Compute both-teams-to-score probabilities from expected goals.
}
