% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalmodel_predict.R
\name{predict_expg}
\alias{predict_expg}
\alias{predict_goals}
\alias{predict_result}
\alias{predict_ou}
\title{Predict Methods for goalmodel Fits}
\usage{
predict_expg(model_fit, team1, team2, x1 = NULL, x2 = NULL, return_df = FALSE)

predict_goals(model_fit, team1, team2, x1 = NULL, x2 = NULL, lwrx = NULL)

predict_result(
  model_fit,
  team1,
  team2,
  x1 = NULL,
  x2 = NULL,
  return_df = FALSE
)

predict_ou(
  model_fit,
  team1,
  team2,
  x1 = NULL,
  x2 = NULL,
  ou = 2.5,
  return_df = FALSE
)
}
\arguments{
\item{model_fit}{A goalmodel object.}

\item{team1}{A character vector with team names, for which to make predictions of.}

\item{team2}{A character vector with team names, for which to make predictions of.}

\item{x1}{Additional covariates to be used for making predictions.}

\item{x2}{Additional covariates to be used for making predictions.}

\item{return_df}{Whether a data.frame should be returned.}

\item{lwrx}{Numeric. The lowest upper limit for the number of goals to compute probabilities for.}

\item{ou}{Numeric, defining the over/under. Default is 2.5.}
}
\description{
Make predictions using a fitted goalmodel.
}
\details{
These functions predict expected goals, the probabiltieis for individual scorelines, or
1x2 results and over/under probabilities.
}
