% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dixoncoles.R
\name{rho.ml}
\alias{rho.ml}
\title{Estimate the Rho Parameter of the Dixon-Coles-Poisson Model}
\usage{
rho.ml(x1, x2, lambda1, lambda2, lower = -0.5, upper = 0.5)
}
\arguments{
\item{x1}{non-negative integer.}

\item{x2}{non-negative integer.}

\item{lambda1}{vector of non-negative parameters of the Dixon-Coles-Poisson distribution.}

\item{lambda2}{vector of non-negative parameters of the Dixon-Coles-Poisson distribution.}

\item{lower}{the lower bound for rho}

\item{upper}{the upper bound for rho}
}
\description{
Given the estimated goals (lambda1, lambda2) and observed goals (x1, x2), estimate the rho parameter of the
the  Dixon-Coles-Poisson model.
}
\section{References}{

\itemize{
 \item{Mark J.  Dixon & Stuart G. Coles (1997) Modelling Association Football Scores and Inefficiencies in the Football Betting Market }
}
}

