% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalmodel_misc.R
\name{summary.goalmodel}
\alias{summary.goalmodel}
\title{Print a summary of a fitted goalmodel.}
\usage{
\method{summary}{goalmodel}(object, ...)
}
\arguments{
\item{object}{A fitted goalmodel object.}

\item{...}{ignored.}
}
\description{
Print a summary of a fitted goalmodel.
}
