% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.r
\name{upsilon.ml}
\alias{upsilon.ml}
\title{Estimate upsilon of the Conway-Maxwell-Poisson Distribution}
\usage{
upsilon.ml(
  x,
  parameters,
  param_type = "lambda",
  method = "sum",
  lower = 0.7,
  upper = 4
)
}
\arguments{
\item{x}{vector of non-negative integers.}

\item{parameters}{vector of non-negative rate parameters (lambda) or mean paramters (mu).}

\item{param_type}{string indicating whether the vector given to the parameters
argument is lambda (default) or mu.}

\item{method}{character; either 'fast' or 'sum'. Passed on to the lambdaCMP function
if param_type = 'mu'.}

\item{lower}{the lower bound for upsilon}

\item{upper}{the upper bound for upsilon}
}
\description{
Given the estimated rate (lambda) or mean (mu) vector, estimate dispersion paramter (uspilon) of
the Conway-Maxwell-Poisson Distribution.
}
\seealso{
\code{\link{CMP}} for more functions and details regarding the Conway-Maxwell-Poisson Distribution.
}
