% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalmodel_misc.R
\name{weights_dc}
\alias{weights_dc}
\title{Compute Dixon-Coles weights}
\usage{
weights_dc(dates, xi = 0, currentDate = NULL)
}
\arguments{
\item{dates}{a vector of dates (of type Date).}

\item{xi}{A numeric with the time dumping factor. Ususally a value buetween 0.001 and 0.003.}

\item{currentDate}{The date which to count backwards from. Default to the latest date in dates.}
}
\description{
Compute Dixon-Coles weights
}
