% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalmodel_misc.R
\name{league_table}
\alias{league_table}
\title{Create league table}
\usage{
league_table(goals1, goals2, team1, team2)
}
\arguments{
\item{goals1}{Numeric, non-negative integer. The number of goals scored by team 1.}

\item{goals2}{Numeric, non-negative integer. The number of goals scored by team 2.}

\item{team1}{Vector of team names.}

\item{team2}{Vector of team names.

#' @return
 A data.frame with one row for each team in the data.}
}
\description{
Creates a league table with basic descriptive statistics and number of points.
}
\details{
Points are calculated as 3 points for win, 1 for draw, and 0 for loss. Results
are sorted by points and ties will be resolved by 1) goal difference and 2) goals scored.
Further ties will not be resolved, and the tied teams will be ranked arbitrarily.

Note that leagues and competitions will often have different rules for how teams
are ranked, especially for how ties are handled. The results provided by this functions
may therefore differ from official league rankings.

Games with missing values in the number of goals scores will be ignored and will
therefore not contribute to the table.
}
