% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalmodel_predict.R
\name{p1x2}
\alias{p1x2}
\title{Compute 1x2 probabilities from expected goals.}
\usage{
p1x2(expg1, expg2, model = "poisson", dispersion = NULL, rho = NULL, uprx = 25)
}
\description{
Compute 1x2 probabilities from expected goals.
}
