% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goalmodel_misc.R
\name{weights_dc}
\alias{weights_dc}
\title{Compute Dixon-Coles weights}
\usage{
weights_dc(dates, xi = 0, currentDate = NULL)
}
\arguments{
\item{dates}{a vector of dates (of type Date).}

\item{xi}{A numeric with the time dumping factor. Ususally a value between 0.001 and 0.003.}

\item{currentDate}{The date which to count backwards from. Default to the latest date in dates.}
}
\description{
This function computes weights that give less weight to games further back in time.
}
\details{
The weigths w are computed using the formula

w = exp(-xi*t)

where xi is a tuning parameter that decides the amount of down-weighting,
and t is the number of days since currentDate. If any date in dates are after
currentDate, the corresponding weights are set to 0.
}
